+++ SERVER/GameServer/Model/User.cs
@@ -12,12 +12,12 @@ namespace GameServer.Model
     public class User
     {
         public NetChannel Channel { get; }
-        public int UserId { get; }
+        public long UserId { get; }
         public string Username { get; }
 
         public Player? Player { get; private set; }
 
-        public User(NetChannel channel, string username, int userId) 
+        public User(NetChannel channel, string username, long userId) 
         {
             Channel = channel;
             Username = username;
